refactor: config, plugins, and walletdb.
