+++ lib/http/base.js
@@ -7,11 +7,19 @@
 
 'use strict';
 
+/* jshint -W069 */
+
 var assert = require('assert');
+var EventEmitter = require('events').EventEmitter;
+var URL = require('url');
+var StringDecoder = require('string_decoder').StringDecoder;
 var AsyncObject = require('../utils/asyncobject');
 var util = require('../utils/util');
-var URL = require('url');
 var co = require('../utils/co');
+var Validator = require('../utils/validator');
+var List = require('../utils/list');
+var fs = require('../utils/fs');
+var crypto = require('../crypto/crypto');
 
 /**
  * HTTPBase
@@ -27,11 +35,14 @@ function HTTPBase(options) {
 
   AsyncObject.call(this);
 
-  this.options = new HTTPBaseOptions(options);
+  this.config = new HTTPBaseOptions(options);
+  this.config.load();
 
   this.server = null;
   this.io = null;
+  this.sockets = new List();
   this.routes = new Routes();
+  this.mounts = [];
   this.stack = [];
   this.hooks = [];
 
@@ -47,13 +58,13 @@ util.inherits(HTTPBase, AsyncObject);
 
 HTTPBase.prototype._init = function _init() {
   var self = this;
-  var backend = this.options.getBackend();
-  var options = this.options.toHTTP();
+  var backend = this.config.getBackend();
+  var options = this.config.toHTTP();
 
   this.server = backend.createServer(options);
 
   this._initRouter();
-  this._initIO();
+  this._initSockets();
 
   this.server.on('connection', function(socket) {
     socket.on('error', function(err) {
@@ -89,13 +100,17 @@ HTTPBase.prototype._init = function _init() {
 
 HTTPBase.prototype._initRouter = function _initRouter() {
   var self = this;
-  this.server.on('request', co(function* (req, res) {
+  this.server.on('request', co(function* (hreq, hres) {
+    var req = new Request(hreq, hres, hreq.url);
+    var res = new Response(hreq, hres);
+
+    req.on('error', nop);
+
     try {
+      req.pause();
       yield self.handleRequest(req, res);
     } catch (e) {
-      if (!res.sent)
-        res.error(e);
-
+      res.error(e.statusCode || 500, e);
       self.emit('error', e);
     }
   }));
@@ -112,14 +127,13 @@ HTTPBase.prototype._initRouter = function _initRouter() {
 HTTPBase.prototype.handleRequest = co(function* handleRequest(req, res) {
   var i, routes, route, params;
 
-  initRequest(req, res, this.options.keyLimit);
+  if (yield this.handleMounts(req, res))
+    return;
 
   this.emit('request', req, res);
 
   if (yield this.handleStack(req, res))
     return;
-
-  req.body = yield this.parseBody(req);
 
   routes = this.routes.getHandlers(req.method);
 
@@ -146,13 +160,136 @@ HTTPBase.prototype.handleRequest = co(function* handleRequest(req, res) {
 });
 
 /**
+ * CORS middleware.
+ * @returns {Function}
+ */
+
+HTTPBase.prototype.cors = function cors() {
+  return co(function* (req, res) {
+    res.setHeader('Access-Control-Allow-Origin', '*');
+    res.setHeader('Access-Control-Allow-Credentials', 'true');
+    res.setHeader(
+      'Access-Control-Allow-Methods',
+      'GET,HEAD,PUT,PATCH,POST,DELETE');
+
+    if (req.method === 'OPTIONS') {
+      res.setStatus(200);
+      res.end();
+      return;
+    }
+  });
+};
+
+/**
+ * Basic auth middleware.
+ * @param {Object} options
+ * @returns {Function}
+ */
+
+HTTPBase.prototype.basicAuth = function basicAuth(options) {
+  var user = options.username;
+  var pass = options.password;
+  var realm = options.realm;
+
+  if (!Buffer.isBuffer(user)) {
+    assert(typeof user === 'string');
+    user = new Buffer(user, 'utf8');
+  }
+
+  if (!Buffer.isBuffer(pass)) {
+    assert(typeof pass === 'string');
+    pass = new Buffer(pass, 'utf8');
+  }
+
+  user = crypto.hash256(user);
+  pass = crypto.hash256(pass);
+
+  // XXX
+  user = null;
+
+  if (!realm)
+    realm = 'server';
+
+  assert(typeof realm === 'string');
+
+  function fail(res) {
+    res.setHeader('WWW-Authenticate', 'Basic realm="' + realm + '"');
+    res.setStatus(401);
+    res.end();
+  }
+
+  return co(function* (req, res) {
+    var auth = req.headers['authorization'];
+    var parts, username, password, digest;
+
+    if (!auth)
+      return fail(res);
+
+    parts = auth.split(' ');
+
+    if (parts.length !== 2)
+      return fail(res);
+
+    if (parts[0] !== 'Basic')
+      return fail(res);
+
+    auth = new Buffer(parts[1], 'base64').toString('utf8');
+    parts = auth.split(':');
+
+    username = parts.shift();
+    password = parts.join(':');
+
+    if (user) {
+      digest = new Buffer(username, 'utf8');
+      digest = crypto.hash256(digest);
+
+      if (!crypto.ccmp(digest, user))
+        return fail(res);
+    }
+
+    digest = new Buffer(password, 'utf8');
+    digest = crypto.hash256(digest);
+
+    if (!crypto.ccmp(digest, pass))
+      return fail(res);
+
+    req.username = username;
+  });
+};
+
+/**
+ * Body parser middleware.
+ * @param {Object} options
+ * @returns {Function}
+ */
+
+HTTPBase.prototype.bodyParser = function bodyParser(options) {
+  var self = this;
+  var opt = new BodyParserOptions(options);
+
+  return co(function* (req, res) {
+    if (req.hasBody)
+      return;
+
+    try {
+      req.resume();
+      req.body = yield self.parseBody(req, opt);
+    } finally {
+      req.pause();
+    }
+
+    req.hasBody = true;
+  });
+};
+
+/**
  * Parse request body.
  * @private
  * @param {ServerRequest} req
  * @returns {Promise}
  */
 
-HTTPBase.prototype.parseBody = co(function* parseBody(req) {
+HTTPBase.prototype.parseBody = co(function* parseBody(req, opt) {
   var body = Object.create(null);
   var type = req.contentType;
   var data;
@@ -160,20 +297,20 @@ HTTPBase.prototype.parseBody = co(function* parseBody(req) {
   if (req.method === 'GET')
     return body;
 
-  data = yield this.readBody(req, 'utf8');
+  data = yield this.readBody(req, 'utf8', opt);
 
   if (!data)
     return body;
 
-  if (this.options.contentType)
-    type = this.options.contentType;
+  if (opt.contentType)
+    type = opt.contentType;
 
   switch (type) {
     case 'json':
       body = JSON.parse(data);
       break;
     case 'form':
-      body = parsePairs(data, this.options.keyLimit);
+      body = parsePairs(data, opt.keyLimit);
       break;
     default:
       break;
@@ -189,10 +326,10 @@ HTTPBase.prototype.parseBody = co(function* parseBody(req) {
  * @returns {Promise}
  */
 
-HTTPBase.prototype.readBody = function readBody(req, enc) {
+HTTPBase.prototype.readBody = function readBody(req, enc, opt) {
   var self = this;
   return new Promise(function(resolve, reject) {
-    return self._readBody(req, enc, resolve, reject);
+    return self._readBody(req, enc, opt, resolve, reject);
   });
 };
 
@@ -205,9 +342,7 @@ HTTPBase.prototype.readBody = function readBody(req, enc) {
  * @param {Function} reject
  */
 
-HTTPBase.prototype._readBody = function _readBody(req, enc, resolve, reject) {
-  var self = this;
-  var StringDecoder = require('string_decoder').StringDecoder;
+HTTPBase.prototype._readBody = function _readBody(req, enc, opt, resolve, reject) {
   var decode = new StringDecoder(enc);
   var hasData = false;
   var total = 0;
@@ -218,7 +353,7 @@ HTTPBase.prototype._readBody = function _readBody(req, enc, resolve, reject) {
     timer = null;
     cleanup();
     reject(new Error('Request body timed out.'));
-  }, 10 * 1000);
+  }, opt.timeout);
 
   function cleanup() {
     req.removeListener('data', onData);
@@ -235,7 +370,7 @@ HTTPBase.prototype._readBody = function _readBody(req, enc, resolve, reject) {
     total += data.length;
     hasData = true;
 
-    if (total > self.options.bodyLimit) {
+    if (total > opt.bodyLimit) {
       reject(new Error('Request body overflow.'));
       return;
     }
@@ -263,6 +398,38 @@ HTTPBase.prototype._readBody = function _readBody(req, enc, resolve, reject) {
   req.on('error', onError);
   req.on('end', onEnd);
 };
+
+/**
+ * Handle mount stack.
+ * @private
+ * @param {HTTPRequest} req
+ * @param {HTTPResponse} res
+ * @returns {Promise}
+ */
+
+HTTPBase.prototype.handleMounts = co(function* handleMounts(req, res) {
+  var url = req.url;
+  var i, route, server;
+
+  for (i = 0; i < this.mounts.length; i++) {
+    route = this.mounts[i];
+    server = route.handler;
+
+    if (!route.hasPrefix(req.pathname))
+      continue;
+
+    assert(url.indexOf(route.path) === 0);
+
+    url = url.substring(route.path.length);
+    req = req.rewrite(url);
+
+    yield server.handleRequest(req, res);
+
+    return true;
+  }
+
+  return false;
+});
 
 /**
  * Handle middleware stack.
@@ -317,11 +484,11 @@ HTTPBase.prototype.handleHooks = co(function* handleHooks(req, res) {
  * @private
  */
 
-HTTPBase.prototype._initIO = function _initIO() {
+HTTPBase.prototype._initSockets = function _initSockets() {
   var self = this;
   var IOServer;
 
-  if (!this.options.sockets)
+  if (!this.config.sockets)
     return;
 
   try {
@@ -339,9 +506,71 @@ HTTPBase.prototype._initIO = function _initIO() {
 
   this.io.attach(this.server);
 
-  this.io.on('connection', function(socket) {
-    self.emit('websocket', socket);
+  this.io.on('connection', function(ws) {
+    self.addSocket(ws);
   });
+};
+
+/**
+ * Broadcast event to channel.
+ * @param {String} channel
+ * @param {String} type
+ * @param {...Object} args
+ */
+
+HTTPBase.prototype.to = function to(channel) {
+  var args = new Array(arguments.length - 1);
+  var i, socket;
+
+  for (i = 1; i < arguments.length; i++)
+    args[i - 1] = arguments[i];
+
+  for (socket = this.sockets.head; socket; socket = socket.next) {
+    if (socket.channels[channel])
+      socket.emit.apply(socket, args);
+  }
+};
+
+/**
+ * Broadcast event to all connections.
+ * @param {String} channel
+ * @param {String} type
+ * @param {...Object} args
+ */
+
+HTTPBase.prototype.all = function all() {
+  var socket;
+
+  for (socket = this.sockets.head; socket; socket = socket.next)
+    socket.emit.apply(socket, arguments);
+};
+
+/**
+ * Initialize websockets.
+ * @private
+ */
+
+HTTPBase.prototype.addSocket = function addSocket(ws) {
+  var self = this;
+  var socket = new WebSocket(ws);
+  var i, route;
+
+  ws.on('error', function(err) {
+    self.emit('error', err);
+  });
+
+  ws.on('disconnect', function() {
+    assert(self.sockets.remove(socket));
+  });
+
+  this.sockets.push(socket);
+
+  for (i = 0; i < this.mounts.length; i++) {
+    route = this.mounts[i];
+    route.handler.addSocket(ws);
+  }
+
+  this.emit('socket', socket);
 };
 
 /**
@@ -351,7 +580,7 @@ HTTPBase.prototype._initIO = function _initIO() {
  */
 
 HTTPBase.prototype._open = function open() {
-  return this.listen(this.options.port, this.options.host);
+  return this.listen(this.config.port, this.config.host);
 };
 
 /**
@@ -381,6 +610,21 @@ HTTPBase.prototype._close = function close() {
 };
 
 /**
+ * Mount a server.
+ * @param {String?} path
+ * @param {HTTPBase} server
+ * @param {Object?} ctx
+ */
+
+HTTPBase.prototype.mount = function mount(path, server, ctx) {
+  if (!server) {
+    server = path;
+    path = null;
+  }
+  this.mounts.push(new Route(ctx || this, path, server));
+};
+
+/**
  * Add a middleware to the stack.
  * @param {String?} path
  * @param {Function} handler
@@ -392,7 +636,7 @@ HTTPBase.prototype.use = function use(path, handler, ctx) {
     handler = path;
     path = null;
   }
-  this.stack.push(new Route(ctx, path, handler));
+  this.stack.push(new Route(ctx || this, path, handler));
 };
 
 /**
@@ -407,7 +651,7 @@ HTTPBase.prototype.hook = function hook(path, handler, ctx) {
     handler = path;
     path = null;
   }
-  this.hooks.push(new Route(ctx, path, handler));
+  this.hooks.push(new Route(ctx || this, path, handler));
 };
 
 /**
@@ -418,7 +662,7 @@ HTTPBase.prototype.hook = function hook(path, handler, ctx) {
  */
 
 HTTPBase.prototype.get = function get(path, handler, ctx) {
-  this.routes.get.push(new Route(ctx, path, handler));
+  this.routes.get.push(new Route(ctx || this, path, handler));
 };
 
 /**
@@ -429,7 +673,7 @@ HTTPBase.prototype.get = function get(path, handler, ctx) {
  */
 
 HTTPBase.prototype.post = function post(path, handler, ctx) {
-  this.routes.post.push(new Route(ctx, path, handler));
+  this.routes.post.push(new Route(ctx || this, path, handler));
 };
 
 /**
@@ -440,7 +684,7 @@ HTTPBase.prototype.post = function post(path, handler, ctx) {
  */
 
 HTTPBase.prototype.put = function put(path, handler, ctx) {
-  this.routes.put.push(new Route(ctx, path, handler));
+  this.routes.put.push(new Route(ctx || this, path, handler));
 };
 
 /**
@@ -451,7 +695,7 @@ HTTPBase.prototype.put = function put(path, handler, ctx) {
  */
 
 HTTPBase.prototype.del = function del(path, handler, ctx) {
-  this.routes.del.push(new Route(ctx, path, handler));
+  this.routes.del.push(new Route(ctx || this, path, handler));
 };
 
 /**
@@ -501,17 +745,14 @@ function HTTPBaseOptions(options) {
 
   this.host = '127.0.0.1';
   this.port = 8080;
-  this.sockets = false;
+  this.sockets = true;
 
   this.ssl = false;
+  this.keyFile = null;
+  this.certFile = null;
   this.key = null;
   this.cert = null;
   this.ca = null;
-
-  this.keyLimit = 100;
-  this.bodyLimit = 20 << 20;
-
-  this.contentType = null;
 
   if (options)
     this.fromOptions(options);
@@ -543,6 +784,28 @@ HTTPBaseOptions.prototype.fromOptions = function fromOptions(options) {
     this.sockets = options.sockets;
   }
 
+  if (options.prefix != null) {
+    assert(typeof options.prefix === 'string');
+    this.prefix = options.prefix;
+    this.keyFile = this.prefix + '/key.pem';
+    this.certFile = this.prefix + '/cert.pem';
+  }
+
+  if (options.ssl != null) {
+    assert(typeof options.ssl === 'boolean');
+    this.ssl = options.ssl;
+  }
+
+  if (options.keyFile != null) {
+    assert(typeof options.keyFile === 'string');
+    this.keyFile = options.keyFile;
+  }
+
+  if (options.certFile != null) {
+    assert(typeof options.certFile === 'string');
+    this.certFile = options.certFile;
+  }
+
   if (options.key != null) {
     assert(typeof options.key === 'string' || Buffer.isBuffer(options.key));
     this.key = options.key;
@@ -558,29 +821,28 @@ HTTPBaseOptions.prototype.fromOptions = function fromOptions(options) {
     this.ca = options.ca;
   }
 
-  if (options.keyLimit != null) {
-    assert(typeof options.keyLimit === 'number');
-    this.keyLimit = options.keyLimit;
-  }
-
-  if (options.bodyLimit != null) {
-    assert(typeof options.bodyLimit === 'number');
-    this.bodyLimit = options.bodyLimit;
-  }
-
-  if (options.ssl != null) {
-    assert(typeof options.ssl === 'boolean');
-    assert(this.key, 'SSL specified with no provided key.');
-    assert(this.cert, 'SSL specified with no provided cert.');
-    this.ssl = options.ssl;
-  }
-
-  if (options.contentType != null) {
-    assert(typeof options.contentType === 'string');
-    this.contentType = options.contentType;
+  if (this.ssl) {
+    assert(this.key || this.keyFile, 'SSL specified with no provided key.');
+    assert(this.cert || this.certFile, 'SSL specified with no provided cert.');
   }
 
   return this;
+};
+
+/**
+ * Load key and cert file.
+ * @private
+ */
+
+HTTPBaseOptions.prototype.load = function load() {
+  if (!this.ssl)
+    return;
+
+  if (this.keyFile)
+    this.key = fs.readFileSync(this.keyFile);
+
+  if (this.certFile)
+    this.cert = fs.readFileSync(this.certFile);
 };
 
 /**
@@ -621,6 +883,54 @@ HTTPBaseOptions.prototype.toHTTP = function toHTTP() {
 };
 
 /**
+ * HTTP Base Options
+ * @alias module:http.BodyParserOptions
+ * @constructor
+ * @param {Object} options
+ */
+
+function BodyParserOptions(options) {
+  if (!(this instanceof BodyParserOptions))
+    return new BodyParserOptions(options);
+
+  this.keyLimit = 100;
+  this.bodyLimit = 20 << 20;
+  this.contentType = null;
+  this.timeout = 10 * 1000;
+
+  if (options)
+    this.fromOptions(options);
+}
+
+/**
+ * Inject properties from object.
+ * @private
+ * @param {Object} options
+ * @returns {BodyParserOptions}
+ */
+
+BodyParserOptions.prototype.fromOptions = function fromOptions(options) {
+  assert(options);
+
+  if (options.keyLimit != null) {
+    assert(typeof options.keyLimit === 'number');
+    this.keyLimit = options.keyLimit;
+  }
+
+  if (options.bodyLimit != null) {
+    assert(typeof options.bodyLimit === 'number');
+    this.bodyLimit = options.bodyLimit;
+  }
+
+  if (options.contentType != null) {
+    assert(typeof options.contentType === 'string');
+    this.contentType = options.contentType;
+  }
+
+  return this;
+};
+
+/**
  * Route
  * @constructor
  * @ignore
@@ -653,7 +963,9 @@ function Route(ctx, path, handler) {
     }
   }
 
-  assert(typeof handler === 'function');
+  assert(handler);
+  assert(typeof handler === 'function' || typeof handler === 'object');
+
   this.handler = handler;
 }
 
@@ -757,205 +1069,77 @@ Routes.prototype.getHandlers = function getHandlers(method) {
   }
 };
 
-/*
- * Helpers
+/**
+ * Request
+ * @constructor
+ * @ignore
  */
 
-function nop() {}
+function Request(req, res, url) {
+  if (!(this instanceof Request))
+    return new Request(req, res, url);
 
-function initRequest(req, res, limit) {
-  var parsed;
+  EventEmitter.call(this);
 
-  req.on('error', nop);
+  this.req = null;
+  this.res = null;
+  this.socket = null;
+  this.method = 'GET';
+  this.headers = Object.create(null);
+  this.contentType = 'bin';
+  this.url = '/';
+  this.pathname = '';
+  this.path = [];
+  this.trailing = false;
+  this.query = Object.create(null);
+  this.params = Object.create(null);
+  this.body = Object.create(null);
+  this.hasBody = false;
+  this.username = null;
+  this.readable = true;
+  this.writable = false;
 
-  assert(req.contentType == null);
-  assert(req.pathname == null);
-  assert(req.path == null);
-  assert(req.query == null);
-  assert(req.params == null);
-  assert(req.body == null);
-
-  req.contentType = parseType(req.headers['content-type']);
-  req.pathname = '';
-  req.path = [];
-  req.query = Object.create(null);
-  req.params = Object.create(null);
-  req.body = Object.create(null);
-
-  assert(req.options == null);
-  assert(req.username == null);
-  assert(req.password == null);
-  assert(req.admin == null);
-  assert(req.wallet == null);
-
-  req.options = Object.create(null);
-  req.username = null;
-  req.password = null;
-  req.admin = false;
-  req.wallet = null;
-
-  assert(res.sent == null);
-  assert(res.send == null);
-  assert(res.error == null);
-  assert(res.redirect == null);
-
-  res.sent = false;
-  res.send = makeSend(res);
-  res.error = makeSendError(req, res);
-  res.redirect = makeRedirect(res);
-
-  parsed = parseURL(req.url, limit);
-
-  req.url = parsed.url;
-  req.pathname = parsed.pathname;
-  req.path = parsed.parts;
-  req.query = parsed.query;
-
-  return parsed;
+  if (req)
+    this.init(req, res, url);
 }
 
-function makeSend(res) {
-  return function send(code, msg, type) {
-    return sendResponse(res, code, msg, type);
-  };
-}
+util.inherits(Request, EventEmitter);
 
-function sendResponse(res, code, msg, type) {
-  var len;
+Request.prototype.init = function init(req, res, url) {
+  var self = this;
 
-  if (res.sent)
-    return;
+  assert(req);
+  assert(res);
 
-  assert(typeof code === 'number', 'Code must be a number.');
+  this.req = req;
+  this.res = res;
+  this.socket = req.socket;
+  this.method = req.method;
+  this.headers = req.headers;
+  this.contentType = parseType(req.headers['content-type']);
 
-  if (msg == null)
-    msg = { error: 'No message.' };
+  req.on('error', function(err) {
+    self.emit('error', err);
+  });
 
-  if (msg && typeof msg === 'object' && !Buffer.isBuffer(msg)) {
-    msg = JSON.stringify(msg, null, 2) + '\n';
-    if (!type)
-      type = 'json';
-    assert(type === 'json', 'Bad type passed with json object.');
-  }
+  req.on('data', function(data) {
+    self.emit('data', data);
+  });
 
-  if (!type)
-    type = typeof msg === 'string' ? 'txt' : 'bin';
+  req.on('end', function() {
+    self.emit('end');
+  });
 
-  res.statusCode = code;
-  res.setHeader('Content-Type', getType(type));
-  res.sent = true;
+  if (url != null)
+    this.parse(url);
+};
 
-  if (typeof msg === 'string') {
-    len = Buffer.byteLength(msg, 'utf8');
-    res.setHeader('Content-Length', len + '');
-    try {
-      res.write(msg, 'utf8');
-      res.end();
-    } catch (e) {
-      ;
-    }
-    return;
-  }
-
-  if (Buffer.isBuffer(msg)) {
-    res.setHeader('Content-Length', msg.length + '');
-    try {
-      res.write(msg);
-      res.end();
-    } catch (e) {
-      ;
-    }
-    return;
-  }
-
-  assert(false, 'Bad object passed to send.');
-}
-
-function makeSendError(req, res) {
-  return function error(err) {
-    return sendError(req, res, err);
-  };
-}
-
-function sendError(req, res, err) {
-  var code, msg;
-
-  if (res.sent)
-    return;
-
-  code = err.statusCode;
-  msg = err.message;
-
-  if (!code)
-    code = 400;
-
-  if (typeof msg !== 'string')
-    msg += '';
-
-  res.send(code, { error: msg });
-
-  try {
-    req.destroy();
-    req.socket.destroy();
-  } catch (e) {
-    ;
-  }
-}
-
-function makeRedirect(res) {
-  return function redirect(code, url) {
-    if (!url) {
-      url = code;
-      code = 301;
-    }
-    res.statusCode = code;
-    res.setHeader('Location', url);
-    res.end();
-  };
-}
-
-function parsePairs(str, limit) {
-  var parts = str.split('&');
-  var data = Object.create(null);
-  var i, index, pair, key, value;
-
-  assert(!limit || parts.length <= limit, 'Too many keys in querystring.');
-
-  for (i = 0; i < parts.length; i++) {
-    pair = parts[i];
-    index = pair.indexOf('=');
-
-    if (index === -1) {
-      key = pair;
-      value = '';
-    } else {
-      key = pair.substring(0, index);
-      value = pair.substring(index + 1);
-    }
-
-    key = unescape(key);
-
-    if (key.length === 0)
-      continue;
-
-    value = unescape(value);
-
-    if (value.length === 0)
-      continue;
-
-    data[key] = value;
-  }
-
-  return data;
-}
-
-function parseURL(str, limit) {
-  var uri = URL.parse(str);
-  var parsed = new ParsedURL(str);
+Request.prototype.parse = function parse(url) {
+  var uri = URL.parse(url);
   var pathname = uri.pathname;
   var query = Object.create(null);
   var trailing = false;
-  var path, parts, url;
+  var path, parts;
 
   if (pathname) {
     pathname = pathname.replace(/\/{2,}/g, '/');
@@ -1006,24 +1190,266 @@ function parseURL(str, limit) {
   }
 
   if (uri.query)
-    query = parsePairs(uri.query, limit);
+    query = parsePairs(uri.query, 100);
 
-  parsed.url = url;
-  parsed.pathname = pathname;
-  parsed.parts = parts;
-  parsed.query = query;
-  parsed.trailing = trailing;
+  this.url = url;
+  this.pathname = pathname;
+  this.path = parts;
+  this.query = query;
+  this.trailing = trailing;
+};
 
-  return parsed;
+Request.prototype.rewrite = function rewrite(url) {
+  var req = new Request();
+  req.init(this.req, this.res, url);
+  req.body = this.body;
+  req.hasBody = this.hasBody;
+  return req;
+};
+
+Request.prototype.valid = function valid() {
+  return new Validator([this.query, this.params, this.body]);
+};
+
+Request.prototype.pipe = function pipe(dest) {
+  return this.req.pipe(dest);
+};
+
+Request.prototype.pause = function pause() {
+  return this.req.pause();
+};
+
+Request.prototype.resume = function resume() {
+  return this.req.resume();
+};
+
+Request.prototype.destroy = function destroy() {
+  return this.req.destroy();
+};
+
+/**
+ * Response
+ * @constructor
+ * @ignore
+ */
+
+function Response(req, res) {
+  if (!(this instanceof Response))
+    return new Response(req, res);
+
+  EventEmitter.call(this);
+
+  this.req = req;
+  this.res = res;
+  this.sent = false;
+  this.readable = false;
+  this.writable = true;
+  this.statusCode = 200;
+  this.res.statusCode = 200;
+
+  if (req)
+    this.init(req, res);
 }
 
-function ParsedURL(original) {
-  this.original = original;
-  this.url = null;
-  this.pathname = null;
-  this.parts = null;
-  this.query = null;
-  this.trailing = false;
+util.inherits(Response, EventEmitter);
+
+Response.prototype.init = function init(req, res) {
+  var self = this;
+
+  assert(req);
+  assert(res);
+
+  res.on('error', function(err) {
+    self.emit('error', err);
+  });
+
+  res.on('drain', function() {
+    self.emit('drain');
+  });
+
+  res.on('close', function() {
+    self.emit('close');
+  });
+};
+
+Response.prototype.setStatus = function setStatus(code) {
+  this.statusCode = code;
+  this.res.statusCode = code;
+};
+
+Response.prototype.setType = function setType(type) {
+  this.setHeader('Content-Type', getType(type));
+};
+
+Response.prototype.hasType = function hasType() {
+  return this.getHeader('Content-Type') != null;
+};
+
+Response.prototype.destroy = function destroy() {
+  return this.res.destroy();
+};
+
+Response.prototype.setHeader = function setHeader(key, value) {
+  return this.res.setHeader(key, value);
+};
+
+Response.prototype.getHeader = function getHeader(key) {
+  return this.res.getHeader(key);
+};
+
+Response.prototype.writeHead = function writeHead(code, headers) {
+  return this.res.writeHead(code, headers);
+};
+
+Response.prototype.write = function write(data, enc) {
+  return this.res.write(data, enc);
+};
+
+Response.prototype.end = function end(data, enc) {
+  this.sent = true;
+  return this.res.end(data, enc);
+};
+
+Response.prototype.error = function error(code, err) {
+  var msg;
+
+  if (this.sent)
+    return;
+
+  msg = err.message;
+
+  if (!code)
+    code = 400;
+
+  if (typeof msg !== 'string')
+    msg += '';
+
+  this.send(code, {
+    error: {
+      type: err.type || 'Error',
+      message: err.stack || msg,
+      code: err.code
+    }
+  });
+
+  try {
+    this.req.destroy();
+    this.req.socket.destroy();
+  } catch (e) {
+    ;
+  }
+};
+
+Response.prototype.redirect = function redirect(code, url) {
+  if (!url) {
+    url = code;
+    code = 301;
+  }
+  this.setStatus(code);
+  this.setHeader('Location', url);
+  this.end();
+};
+
+Response.prototype.send = function send(code, msg, type) {
+  var len;
+
+  if (this.sent)
+    return;
+
+  assert(typeof code === 'number', 'Code must be a number.');
+
+  if (msg == null) {
+    msg = {
+      error: {
+        type: 'Error',
+        message: 'No message.'
+      }
+    };
+  }
+
+  if (msg && typeof msg === 'object' && !Buffer.isBuffer(msg)) {
+    msg = JSON.stringify(msg, null, 2) + '\n';
+    if (!type)
+      type = 'json';
+    assert(type === 'json', 'Bad type passed with json object.');
+  }
+
+  if (!type && !this.hasType())
+    type = typeof msg === 'string' ? 'txt' : 'bin';
+
+  this.setStatus(code);
+
+  if (type)
+    this.setType(type);
+
+  if (typeof msg === 'string') {
+    len = Buffer.byteLength(msg, 'utf8');
+    this.setHeader('Content-Length', len + '');
+    try {
+      this.write(msg, 'utf8');
+      this.end();
+    } catch (e) {
+      ;
+    }
+    return;
+  }
+
+  if (Buffer.isBuffer(msg)) {
+    this.setHeader('Content-Length', msg.length + '');
+    try {
+      this.write(msg);
+      this.end();
+    } catch (e) {
+      ;
+    }
+    return;
+  }
+
+  assert(false, 'Bad object passed to send.');
+};
+
+/*
+ * Helpers
+ */
+
+function nop() {}
+
+function parsePairs(str, limit) {
+  var parts = str.split('&');
+  var data = Object.create(null);
+  var i, index, pair, key, value;
+
+  if (parts.length > limit)
+    return data;
+
+  assert(!limit || parts.length <= limit, 'Too many keys in querystring.');
+
+  for (i = 0; i < parts.length; i++) {
+    pair = parts[i];
+    index = pair.indexOf('=');
+
+    if (index === -1) {
+      key = pair;
+      value = '';
+    } else {
+      key = pair.substring(0, index);
+      value = pair.substring(index + 1);
+    }
+
+    key = unescape(key);
+
+    if (key.length === 0)
+      continue;
+
+    value = unescape(value);
+
+    if (value.length === 0)
+      continue;
+
+    data[key] = value;
+  }
+
+  return data;
 }
 
 function unescape(str) {
@@ -1041,6 +1467,8 @@ function getType(type) {
       return 'application/x-www-form-urlencoded; charset=utf-8';
     case 'html':
       return 'text/html; charset=utf-8';
+    case 'xml':
+      return 'application/xml; charset=utf-8';
     case 'js':
       return 'application/javascript; charset=utf-8';
     case 'css':
@@ -1050,7 +1478,7 @@ function getType(type) {
     case 'bin':
       return 'application/octet-stream';
     default:
-      throw new Error('Unknown type: ' + type);
+      return type;
   }
 }
 
@@ -1082,6 +1510,119 @@ function parseType(type) {
       return 'bin';
   }
 }
+
+/**
+ * WebSocket
+ * @constructor
+ * @ignore
+ * @param {SocketIO.Socket}
+ */
+
+function WebSocket(socket) {
+  if (!(this instanceof WebSocket))
+    return new WebSocket(socket);
+
+  EventEmitter.call(this);
+
+  this.socket = socket;
+  this.remoteAddress = socket.conn.remoteAddress;
+  this.hooks = {};
+  this.channels = {};
+  this.auth = false;
+  this.prev = null;
+  this.next = null;
+
+  this.init();
+}
+
+util.inherits(WebSocket, EventEmitter);
+
+WebSocket.prototype.init = function init() {
+  var self = this;
+  var socket = this.socket;
+  var onevent = socket.onevent.bind(socket);
+
+  socket.onevent = function(packet) {
+    var result = onevent(packet);
+    self.onevent(packet);
+    return result;
+  };
+
+  socket.on('error', function(err) {
+    self.dispatch('error', err);
+  });
+
+  socket.on('disconnect', function() {
+    self.dispatch('disconnect');
+  });
+};
+
+WebSocket.prototype.onevent = co(function* onevent(packet) {
+  var socket = this.socket;
+  var args = (packet.data || []).slice();
+  var type = args.shift() || '';
+  var ack, result;
+
+  if (typeof args[args.length - 1] === 'function')
+    ack = args.pop();
+  else
+    ack = this.socket.ack(packet.id);
+
+  this.dispatch(type, args);
+
+  try {
+    result = yield this.fire(type, args);
+  } catch (e) {
+    ack({
+      type: e.type || 'Error',
+      message: e.stack,
+      code: e.code
+    });
+    return;
+  }
+
+  if (result === undefined)
+    return;
+
+  ack(null, result);
+});
+
+WebSocket.prototype.hook = function hook(type, handler) {
+  // assert(!this.hooks[type], 'Event already added.');
+  if (this.hooks[type])
+    return;
+  this.hooks[type] = handler;
+};
+
+WebSocket.prototype.fire = co(function* fire(type, args) {
+  var handler = this.hooks[type];
+
+  if (!handler)
+    return;
+
+  return yield handler(args);
+});
+
+WebSocket.prototype.join = function join(channel) {
+  this.channels[channel] = true;
+};
+
+WebSocket.prototype.leave = function leave(channel) {
+  delete this.channels[channel];
+};
+
+WebSocket.prototype.dispatch = function dispatch() {
+  var emit = EventEmitter.prototype.emit;
+  return emit.apply(this, arguments);
+};
+
+WebSocket.prototype.emit = function emit() {
+  return this.socket.emit.apply(this.socket, arguments);
+};
+
+WebSocket.prototype.disconnect = function disconnect() {
+  return this.socket.disconnect();
+};
 
 /*
  * Expose
