+++ lib/http/client.js
@@ -108,8 +108,8 @@ HTTPClient.prototype._open = co(function* _open() {
     self.emit('balance', balance);
   });
 
-  yield this._onConnect();
-  yield this._sendAuth();
+  yield this.onConnect();
+  yield this.sendAuth();
 });
 
 /**
@@ -134,7 +134,7 @@ HTTPClient.prototype._close = function close() {
  * @returns {Promise}
  */
 
-HTTPClient.prototype._onConnect = function _onConnect() {
+HTTPClient.prototype.onConnect = function onConnect() {
   var self = this;
   return new Promise(function(resolve, reject) {
     self.socket.once('connect', resolve);
@@ -147,12 +147,29 @@ HTTPClient.prototype._onConnect = function _onConnect() {
  * @returns {Promise}
  */
 
-HTTPClient.prototype._sendAuth = function _sendAuth() {
+HTTPClient.prototype.sendAuth = function sendAuth() {
   var self = this;
   return new Promise(function(resolve, reject) {
     self.socket.emit('auth', self.apiKey, function(err) {
       if (err)
-        return reject(new Error(err.error));
+        return reject(new Error(err.message));
+      resolve();
+    });
+  });
+};
+
+/**
+ * Wait for websocket auth.
+ * @private
+ * @returns {Promise}
+ */
+
+HTTPClient.prototype.sendWalletAuth = function sendWalletAuth() {
+  var self = this;
+  return new Promise(function(resolve, reject) {
+    self.socket.emit('wallet auth', self.apiKey, function(err) {
+      if (err)
+        return reject(new Error(err.message));
       resolve();
     });
   });
@@ -220,7 +237,7 @@ HTTPClient.prototype._request = co(function* _request(method, endpoint, json) {
 
   if (res.statusCode !== 200) {
     if (res.body.error)
-      throw new Error(res.body.error);
+      throw new Error(res.body.error.message);
     throw new Error('Status code: ' + res.statusCode);
   }
 
@@ -369,7 +386,7 @@ HTTPClient.prototype.broadcast = function broadcast(tx) {
 
 HTTPClient.prototype.rescan = function rescan(height) {
   var options = { height: height };
-  return this._post('/rescan', options);
+  return this._post('/_admin/rescan', options);
 };
 
 /**
@@ -380,7 +397,7 @@ HTTPClient.prototype.rescan = function rescan(height) {
 
 HTTPClient.prototype.reset = function reset(block) {
   var options = { block: block };
-  return this._post('/reset', options);
+  return this._post('/_admin/reset', options);
 };
 
 /**
@@ -389,7 +406,7 @@ HTTPClient.prototype.reset = function reset(block) {
  */
 
 HTTPClient.prototype.resend = function resend() {
-  return this._post('/resend', {});
+  return this._post('/_admin/resend', {});
 };
 
 /**
@@ -400,7 +417,7 @@ HTTPClient.prototype.resend = function resend() {
 
 HTTPClient.prototype.backup = function backup(path) {
   var options = { path: path };
-  return this._post('/backup', options);
+  return this._post('/_admin/backup', options);
 };
 
 /**
@@ -417,7 +434,7 @@ HTTPClient.prototype.join = function join(id, token) {
   return new Promise(function(resolve, reject) {
     self.socket.emit('wallet join', id, token, function(err) {
       if (err)
-        return reject(new Error(err.error));
+        return reject(new Error(err.message));
       resolve();
     });
   });
@@ -437,7 +454,7 @@ HTTPClient.prototype.leave = function leave(id) {
   return new Promise(function(resolve, reject) {
     self.socket.emit('wallet leave', id, function(err) {
       if (err)
-        return reject(new Error(err.error));
+        return reject(new Error(err.message));
       resolve();
     });
   });
@@ -475,7 +492,7 @@ HTTPClient.prototype.getWallets = function getWallets() {
  */
 
 HTTPClient.prototype.createWallet = function createWallet(options) {
-  return this._post('/wallet', options);
+  return this._put('/wallet/' + options.id, options);
 };
 
 /**
@@ -934,9 +951,9 @@ HTTPClient.prototype.createAccount = function createAccount(id, options) {
   if (typeof options === 'string')
     options = { account: options };
 
-  path = '/wallet/' + id + '/account';
+  path = '/wallet/' + id + '/account' + options.account;
 
-  return this._post(path, options);
+  return this._put(path, options);
 };
 
 /**
