+++ lib/http/wallet.js
@@ -115,6 +115,7 @@ HTTPWallet.prototype.open = co(function* open(options) {
   assert(this.id, 'No ID provided.');
 
   yield this.client.open();
+  yield this.client.sendWalletAuth();
   yield this.client.join(this.id, this.token);
 });
 
@@ -128,6 +129,7 @@ HTTPWallet.prototype.create = co(function* create(options) {
   var wallet;
 
   yield this.client.open();
+  yield this.client.sendWalletAuth();
 
   wallet = yield this.client.createWallet(options);
