+++ lib/mining/miner.js
@@ -107,6 +107,9 @@ Miner.prototype._open = co(function* open() {
 
   this.logger.info('Miner loaded (flags=%s).',
     this.options.coinbaseFlags.toString('utf8'));
+
+  if (this.addresses.length === 0)
+    this.logger.warning('No reward address is set for miner!');
 });
 
 /**
@@ -387,7 +390,8 @@ Miner.prototype.addAddress = function addAddress(address) {
  */
 
 Miner.prototype.getAddress = function getAddress() {
-  assert(this.addresses.length !== 0, 'No address passed in for miner.');
+  if (this.addresses.length === 0)
+    return;
   return this.addresses[Math.random() * this.addresses.length | 0];
 };
