+++ lib/mining/minerblock.js
@@ -14,6 +14,7 @@ var util = require('../utils/util');
 var co = require('../utils/co');
 var StaticWriter = require('../utils/staticwriter');
 var Network = require('../protocol/network');
+var Address = require('../primitives/address');
 var TX = require('../primitives/tx');
 var Block = require('../primitives/block');
 var Input = require('../primitives/input');
@@ -232,8 +233,10 @@ MinerBlock.prototype._init = function _init() {
   input.script.compile();
 
   // Setup output script (variable size).
-  output.script.clear();
-  output.script.fromAddress(this.address);
+  if (this.address) {
+    output.script.clear();
+    output.script.fromAddress(this.address);
+  }
 
   // Update commitments.
   this.refresh();
@@ -343,6 +346,23 @@ MinerBlock.prototype.extraNonce = function extraNonce() {
   bw.writeU32BE(this.nonce1);
   bw.writeU32BE(this.nonce2);
   return bw.render();
+};
+
+/**
+ * Set the reward output address.
+ * @param {Address} address
+ */
+
+MinerBlock.prototype.setAddress = function setAddress(address) {
+  var output = this.coinbase.outputs[0];
+
+  this.address = Address(address);
+
+  output.script.clear();
+  output.script.fromAddress(this.address);
+
+  // Update commitments.
+  this.refresh();
 };
 
 /**
