+++ lib/net/pool.js
@@ -268,7 +268,6 @@ Pool.prototype.resetChain = function resetChain() {
 
 Pool.prototype._close = co(function* close() {
   yield this.disconnect();
-  yield this.hosts.close();
 });
 
 /**
