+++ lib/node/config.js
@@ -243,8 +243,8 @@ Config.prototype.str = function str(key, fallback) {
   if (value === null)
     return fallback;
 
-  assert(typeof value === 'string',
-    'Passed in config option is of wrong type.');
+  if (typeof value !== 'string')
+    throw new Error(key + ' must be a string.');
 
   return value;
 };
@@ -266,15 +266,18 @@ Config.prototype.num = function num(key, fallback) {
     return fallback;
 
   if (typeof value !== 'string') {
-    assert(typeof value === 'number',
-      'Passed in config option is of wrong type.');
+    if (typeof value !== 'number')
+      throw new Error(key + ' must be a string.');
     return value;
   }
+
+  if (!/^\d+$/.test(value))
+    throw new Error(key + ' must be a number.');
 
   value = parseInt(value, 10);
 
   if (!isFinite(value))
-    return fallback;
+    throw new Error(key + ' must be a number.');
 
   return value;
 };
@@ -296,8 +299,8 @@ Config.prototype.bool = function bool(key, fallback) {
     return fallback;
 
   if (typeof value !== 'string') {
-    assert(typeof value === 'boolean',
-      'Passed in config option is of wrong type.');
+    if (typeof value !== 'boolean')
+      throw new Error(key + ' must be a boolean.');
     return value;
   }
 
@@ -307,7 +310,7 @@ Config.prototype.bool = function bool(key, fallback) {
   if (value === 'false' || value === '0')
     return false;
 
-  return fallback;
+  throw new Error(key + ' must be a boolean.');
 };
 
 /**
@@ -319,6 +322,7 @@ Config.prototype.bool = function bool(key, fallback) {
 
 Config.prototype.buf = function buf(key, fallback) {
   var value = this.get(key);
+  var data;
 
   if (fallback === undefined)
     fallback = null;
@@ -327,12 +331,17 @@ Config.prototype.buf = function buf(key, fallback) {
     return fallback;
 
   if (typeof value !== 'string') {
-    assert(Buffer.isBuffer(value),
-      'Passed in config option is of wrong type.');
+    if (!Buffer.isBuffer(value))
+      throw new Error(key + ' must be a buffer.');
     return value;
   }
 
-  return new Buffer(value, 'hex');
+  data = new Buffer(value, 'hex');
+
+  if (data.length !== value.length / 2)
+    throw new Error(key + ' must be a hex string.');
+
+  return data;
 };
 
 /**
@@ -352,8 +361,8 @@ Config.prototype.list = function list(key, fallback) {
     return fallback;
 
   if (typeof value !== 'string') {
-    assert(Array.isArray(value),
-      'Passed in config option is of wrong type.');
+    if (!Array.isArray(value))
+      throw new Error(key + ' must be an array.');
     return value;
   }
 
@@ -377,8 +386,8 @@ Config.prototype.obj = function obj(key, fallback) {
     return fallback;
 
   if (typeof value !== 'string') {
-    assert(value && typeof value === 'object',
-      'Passed in config option is of wrong type.');
+    if (!value || typeof value !== 'object')
+      throw new Error(key + ' must be an object.');
     return value;
   }
 
@@ -388,8 +397,8 @@ Config.prototype.obj = function obj(key, fallback) {
     ;
   }
 
-  assert(value && typeof value === 'object',
-    'Passed in config option is of wrong type.');
+  if (!value || typeof value !== 'object')
+    throw new Error(key + ' must be an object.');
 
   return value;
 };
@@ -411,13 +420,12 @@ Config.prototype.func = function func(key, fallback) {
     return fallback;
 
   if (typeof value !== 'string') {
-    assert(value && typeof value === 'function',
-      'Passed in config option is of wrong type.');
+    if (!value || typeof value !== 'function')
+      throw new Error(key + ' must be a function.');
     return value;
   }
 
-  assert(false,
-    'Passed in config option is of wrong type.');
+  throw new Error(key + ' must be a function.');
 };
 
 /**
@@ -486,8 +494,8 @@ Config.prototype.boolpath = function boolpath(key, fallback) {
     return fallback;
 
   if (typeof value !== 'string') {
-    assert(typeof value === 'boolean' || typeof value === 'string',
-      'Passed in config option is of wrong type.');
+    if (typeof value !== 'boolean' && typeof value !== 'string')
+      throw new Error(key + ' must be a boolean or string.');
     return value;
   }
