+++ lib/node/fullnode.js
@@ -15,12 +15,11 @@ var Fees = require('../mempool/fees');
 var Mempool = require('../mempool/mempool');
 var Pool = require('../net/pool');
 var Miner = require('../mining/miner');
-var WalletDB = require('../wallet/walletdb');
 var HTTPServer = require('../http/server');
 
 /**
  * Respresents a fullnode complete with a
- * chain, mempool, miner, wallet, etc.
+ * chain, mempool, miner, etc.
  * @alias module:node.FullNode
  * @extends Node
  * @constructor
@@ -30,7 +29,6 @@ var HTTPServer = require('../http/server');
  * @property {Mempool} mempool
  * @property {Pool} pool
  * @property {Miner} miner
- * @property {WalletDB} walletdb
  * @property {HTTPServer} http
  * @emits FullNode#block
  * @emits FullNode#tx
@@ -126,22 +124,6 @@ function FullNode(options) {
     reservedSigops: this.config.num('reserved-sigops')
   });
 
-  // Wallet database needs access to fees.
-  this.walletdb = new WalletDB({
-    network: this.network,
-    logger: this.logger,
-    client: this.client,
-    db: this.config.str('db'),
-    prefix: this.config.prefix,
-    maxFiles: this.config.num('walletdb-max-files'),
-    cacheSize: this.config.mb('walletdb-cache-size'),
-    witness: false,
-    checkpoints: this.config.bool('checkpoints'),
-    startHeight: this.config.num('start-height'),
-    wipeNoReally: this.config.bool('wipe-no-really'),
-    verify: false
-  });
-
   // HTTP needs access to the node.
   if (!HTTPServer.unsupported) {
     this.http = new HTTPServer({
@@ -180,7 +162,6 @@ FullNode.prototype._init = function _init() {
   this.mempool.on('error', onError);
   this.pool.on('error', onError);
   this.miner.on('error', onError);
-  this.walletdb.on('error', onError);
 
   if (this.http)
     this.http.on('error', onError);
@@ -217,6 +198,8 @@ FullNode.prototype._init = function _init() {
     }
     self.emit('reset', tip);
   }));
+
+  this.loadPlugins();
 };
 
 /**
@@ -231,10 +214,8 @@ FullNode.prototype._open = co(function* open() {
   yield this.mempool.open();
   yield this.miner.open();
   yield this.pool.open();
-  yield this.walletdb.open();
 
-  // Ensure primary wallet.
-  yield this.openWallet();
+  yield this.openPlugins();
 
   if (this.http)
     yield this.http.open();
@@ -252,11 +233,8 @@ FullNode.prototype._close = co(function* close() {
   if (this.http)
     yield this.http.close();
 
-  yield this.wallet.destroy();
+  yield this.closePlugins();
 
-  this.wallet = null;
-
-  yield this.walletdb.close();
   yield this.pool.close();
   yield this.miner.close();
   yield this.mempool.close();
