+++ lib/node/node.js
@@ -13,7 +13,6 @@ var util = require('../utils/util');
 var co = require('../utils/co');
 var Network = require('../protocol/network');
 var Logger = require('./logger');
-var NodeClient = require('./nodeclient');
 var workerPool = require('../workers/workerpool').pool;
 var ec = require('../crypto/ec');
 var native = require('../utils/native');
@@ -48,8 +47,6 @@ function Node(options) {
   this.mempool = null;
   this.pool = null;
   this.miner = null;
-  this.walletdb = null;
-  this.wallet = null;
   this.http = null;
   this.client = null;
 
@@ -90,13 +87,13 @@ Node.prototype.init = function init() {
   var self = this;
 
   this.initOptions();
-  this.loadPlugins();
-
-  // Local client for walletdb
-  this.client = new NodeClient(this);
 
   this.hook('preopen', function() {
     return self.handlePreopen();
+  });
+
+  this.hook('preopen', function() {
+    return self.handlePreclose();
   });
 
   this.hook('open', function() {
@@ -117,6 +114,7 @@ Node.prototype.handlePreopen = co(function* handlePreopen() {
   var self = this;
 
   yield fs.mkdirp(this.config.prefix);
+
   yield this.logger.open();
 
   this.bind(this.network.time, 'offset', function(offset) {
@@ -173,8 +171,14 @@ Node.prototype.handleOpen = co(function* handleOpen() {
     this.logger.warning('Warning: worker pool is disabled.');
     this.logger.warning('Verification will be slow.');
   }
+});
 
-  yield this.openPlugins();
+/**
+ * Open node. Bind all events.
+ * @private
+ */
+
+Node.prototype.handlePreclose = co(function* handlePreclose() {
 });
 
 /**
@@ -184,8 +188,6 @@ Node.prototype.handleOpen = co(function* handleOpen() {
 
 Node.prototype.handleClose = co(function* handleClose() {
   var i, bound;
-
-  yield this.closePlugins();
 
   this.startTime = -1;
 
@@ -265,36 +267,6 @@ Node.prototype.uptime = function uptime() {
 };
 
 /**
- * Open and ensure primary wallet.
- * @returns {Promise}
- */
-
-Node.prototype.openWallet = co(function* openWallet() {
-  var options, wallet;
-
-  assert(!this.wallet);
-
-  options = {
-    id: 'primary',
-    passphrase: this.config.str('passphrase')
-  };
-
-  wallet = yield this.walletdb.ensure(options);
-
-  this.logger.info(
-    'Loaded wallet with id=%s wid=%d address=%s',
-    wallet.id, wallet.wid, wallet.getAddress());
-
-  if (this.miner && this.miner.addresses.length === 0)
-    this.miner.addAddress(wallet.getAddress());
-
-  this.wallet = wallet;
-
-  if (this.http && this.http.rpc && !this.http.rpc.wallet)
-    this.http.rpc.wallet = wallet;
-});
-
-/**
  * Attach a plugin.
  * @param {Object} plugin
  */
@@ -312,25 +284,24 @@ Node.prototype.use = function use(plugin) {
   assert(typeof instance.open === 'function', '`open` must be a function.');
   assert(typeof instance.close === 'function', '`close` must be a function.');
 
-  if (plugin.name) {
-    assert(typeof plugin.name === 'string', '`name` must be a string.');
+  if (plugin.id) {
+    assert(typeof plugin.id === 'string', '`name` must be a string.');
 
     // Reserved names
-    switch (plugin.name) {
+    switch (plugin.id) {
       case 'logger':
       case 'chain':
       case 'fees':
       case 'mempool':
       case 'miner':
       case 'pool':
-      case 'walletdb':
-        assert(false, plugin.name + ' is already added.');
+        assert(false, plugin.id + ' is already added.');
         break;
     }
 
-    assert(!this.plugins[plugin.name], plugin.name + ' is already added.');
+    assert(!this.plugins[plugin.id], plugin.id + ' is already added.');
 
-    this.plugins[plugin.name] = instance;
+    this.plugins[plugin.id] = instance;
   }
 
   this.stack.push(instance);
@@ -366,9 +337,6 @@ Node.prototype.require = function require(name) {
     case 'pool':
       assert(this.pool, 'pool is not loaded.');
       return this.pool;
-    case 'walletdb':
-      assert(this.walletdb, 'walletdb is not loaded.');
-      return this.walletdb;
   }
 
   plugin = this.plugins[name];
@@ -392,9 +360,16 @@ Node.prototype.loadPlugins = function loadPlugins() {
 
   for (i = 0; i < plugins.length; i++) {
     name = plugins[i];
+
     assert(typeof name === 'string',
       'Plugin name must be a string.');
+
+    // Temporary until we separate walletdb out.
+    if (name === 'walletdb')
+      name = __dirname + '/../wallet/walletdb';
+
     plugin = loader(name);
+
     this.use(plugin);
   }
 };
