+++ lib/node/spvnode.js
@@ -13,12 +13,11 @@ var Lock = require('../utils/lock');
 var Node = require('./node');
 var Chain = require('../blockchain/chain');
 var Pool = require('../net/pool');
-var WalletDB = require('../wallet/walletdb');
 var HTTPServer = require('../http/server');
 
 /**
  * Create an spv node which only maintains
- * a chain, a pool, and a wallet database.
+ * a chain, a pool, and an http server.
  * @alias module:node.SPVNode
  * @extends Node
  * @constructor
@@ -27,11 +26,9 @@ var HTTPServer = require('../http/server');
  * @param {Buffer?} options.sslCert
  * @param {Number?} options.httpPort
  * @param {String?} options.httpHost
- * @param {Object?} options.wallet - Primary {@link Wallet} options.
  * @property {Boolean} loaded
  * @property {Chain} chain
  * @property {Pool} pool
- * @property {WalletDB} walletdb
  * @property {HTTPServer} http
  * @emits SPVNode#block
  * @emits SPVNode#tx
@@ -74,22 +71,6 @@ function SPVNode(options) {
     listen: false
   });
 
-  this.walletdb = new WalletDB({
-    network: this.network,
-    logger: this.logger,
-    client: this.client,
-    db: this.config.str('db'),
-    prefix: this.config.prefix,
-    maxFiles: this.config.num('max-files'),
-    cacheSize: this.config.mb('cache-size'),
-    witness: false,
-    checkpoints: this.config.bool('checkpoints'),
-    startHeight: this.config.num('start-height'),
-    wipeNoReally: this.config.bool('wipe-no-really'),
-    verify: true,
-    spv: true
-  });
-
   if (!HTTPServer.unsupported) {
     this.http = new HTTPServer({
       network: this.network,
@@ -129,7 +110,6 @@ SPVNode.prototype._init = function _init() {
   // Bind to errors
   this.chain.on('error', onError);
   this.pool.on('error', onError);
-  this.walletdb.on('error', onError);
 
   if (this.http)
     this.http.on('error', onError);
@@ -165,6 +145,8 @@ SPVNode.prototype._init = function _init() {
   this.chain.on('reset', function(tip) {
     self.emit('reset', tip);
   });
+
+  this.loadPlugins();
 };
 
 /**
@@ -177,10 +159,8 @@ SPVNode.prototype._init = function _init() {
 SPVNode.prototype._open = co(function* open(callback) {
   yield this.chain.open();
   yield this.pool.open();
-  yield this.walletdb.open();
 
-  // Ensure primary wallet.
-  yield this.openWallet();
+  yield this.openPlugins();
 
   if (this.http)
     yield this.http.open();
@@ -198,11 +178,8 @@ SPVNode.prototype._close = co(function* close() {
   if (this.http)
     yield this.http.close();
 
-  yield this.wallet.destroy();
+  yield this.closePlugins();
 
-  this.wallet = null;
-
-  yield this.walletdb.close();
   yield this.pool.close();
   yield this.chain.close();
 });
