+++ lib/utils/index.js
@@ -31,3 +31,4 @@ exports.BufferReader = require('./reader');
 exports.StaticWriter = require('./staticwriter');
 exports.util = require('./util');
 exports.BufferWriter = require('./writer');
+exports.Validator = require('./validator');
