+++ lib/utils/util.js
@@ -1037,3 +1037,30 @@ util.promisify = function promisify(func) {
     return Promise.resolve(result);
   };
 };
+
+/**
+ * Get memory usage info.
+ * @returns {Object}
+ */
+
+util.memoryUsage = function memoryUsage() {
+  var mem;
+
+  if (!process.memoryUsage) {
+    return {
+      total: 0,
+      jsHeap: 0,
+      jsHeapTotal: 0,
+      nativeHeap: 0
+    };
+  }
+
+  mem = process.memoryUsage();
+
+  return {
+    total: util.mb(mem.rss),
+    jsHeap: util.mb(mem.heapUsed),
+    jsHeapTotal: util.mb(mem.heapTotal),
+    nativeHeap: util.mb(mem.rss - mem.heapTotal)
+  };
+};
