+++ lib/wallet/common.js
@@ -23,6 +23,9 @@ common.isName = function isName(key) {
   if (typeof key !== 'string')
     return false;
 
+  if (key === '_admin')
+    return false;
+
   if (key === '__proto__')
     return false;
