+++ lib/wallet/wallet.js
@@ -1424,7 +1424,7 @@ Wallet.prototype._fund = co(function* fund(mtx, options) {
   rate = options.rate;
 
   if (rate == null)
-    rate = yield this.db.estimateFee();
+    rate = yield this.db.estimateFee(options.blocks);
 
   if (options.smart) {
     coins = yield this.getSmartCoins(options.account);
