+++ lib/wallet/walletdb.js
@@ -26,6 +26,8 @@ var Logger = require('../node/logger');
 var Outpoint = require('../primitives/outpoint');
 var layouts = require('./layout');
 var records = require('./records');
+var NodeClient = require('../node/nodeclient');
+var HTTP = require('./http');
 var layout = layouts.walletdb;
 var ChainState = records.ChainState;
 var BlockMapRecord = records.BlockMapRecord;
@@ -62,6 +64,20 @@ function WalletDB(options) {
   this.logger = this.options.logger;
   this.client = this.options.client;
   this.db = LDB(this.options);
+  this.primary = null;
+
+  this.http = new HTTP({
+    walletdb: this,
+    network: this.network,
+    logger: this.logger,
+    prefix: this.options.prefix,
+    apiKey: this.options.apiKey,
+    walletAuth: this.options.walletAuth,
+    noAuth: this.options.noAuth,
+    host: this.options.host,
+    port: this.options.port,
+    ssl: this.options.ssl
+  });
 
   this.state = new ChainState();
   this.wallets = Object.create(null);
@@ -82,6 +98,54 @@ function WalletDB(options) {
 }
 
 util.inherits(WalletDB, AsyncObject);
+
+/**
+ * Plugin name.
+ * @const {String}
+ */
+
+WalletDB.id = 'walletdb';
+
+/**
+ * Plugin initialization.
+ * @param {Node} node
+ * @returns {WalletDB}
+ */
+
+WalletDB.init = function init(node) {
+  var config = node.config;
+  var client = new NodeClient(node);
+  var wdb;
+
+  wdb = new WalletDB({
+    network: node.network,
+    logger: node.logger,
+    client: client,
+    prefix: config.prefix,
+    db: config.str(['wallet-db', 'db']),
+    maxFiles: config.num('wallet-max-files'),
+    cacheSize: config.mb('wallet-cache-size'),
+    witness: config.bool('wallet-witness'),
+    checkpoints: config.bool('wallet-checkpoints'),
+    startHeight: config.num('wallet-start-height'),
+    wipeNoReally: config.bool('wallet-wipe-no-really'),
+    apiKey: config.str(['wallet-api-key', 'api-key']),
+    walletAuth: config.bool('wallet-auth'),
+    noAuth: config.bool('no-auth'),
+    ssl: config.str('wallet-ssl'),
+    host: config.str('wallet-host'),
+    port: config.num('wallet-port'),
+    spv: node.chain.options.spv,
+    verify: node.chain.options.spv
+  });
+
+  if (node.http) {
+    wdb.http.attach(node.http);
+    wdb.http.rpc.attach(node.http.rpc);
+  }
+
+  return wdb;
+};
 
 /**
  * Database layout.
@@ -119,6 +183,8 @@ WalletDB.prototype._init = function _init() {
  */
 
 WalletDB.prototype._open = co(function* open() {
+  var wallet;
+
   yield this.db.open();
   yield this.db.checkVersion('V', 6);
 
@@ -134,6 +200,17 @@ WalletDB.prototype._open = co(function* open() {
     this.depth,
     this.state.height,
     this.state.startHeight);
+
+  wallet = yield this.ensure({
+    id: 'primary'
+  });
+
+  this.logger.info(
+    'Loaded wallet with id=%s wid=%d address=%s',
+    wallet.id, wallet.wid, wallet.getAddress());
+
+  this.primary = wallet;
+  this.http.rpc.wallet = wallet;
 });
 
 /**
@@ -2185,6 +2262,12 @@ function WalletOptions(options) {
   this.startHeight = 0;
   this.keepBlocks = this.network.block.keepBlocks;
   this.wipeNoReally = false;
+  this.apiKey = null;
+  this.walletAuth = false;
+  this.noAuth = false;
+  this.ssl = false;
+  this.host = '127.0.0.1';
+  this.port = this.network.rpcPort + 2;
 
   if (options)
     this.fromOptions(options);
@@ -2201,6 +2284,7 @@ WalletOptions.prototype.fromOptions = function fromOptions(options) {
   if (options.network != null) {
     this.network = Network.get(options.network);
     this.keepBlocks = this.network.block.keepBlocks;
+    this.port = this.network.rpcPort + 2;
   }
 
   if (options.logger != null) {
@@ -2268,6 +2352,36 @@ WalletOptions.prototype.fromOptions = function fromOptions(options) {
   if (options.wipeNoReally != null) {
     assert(typeof options.wipeNoReally === 'boolean');
     this.wipeNoReally = options.wipeNoReally;
+  }
+
+  if (options.apiKey != null) {
+    assert(typeof options.apiKey === 'string');
+    this.apiKey = options.apiKey;
+  }
+
+  if (options.walletAuth != null) {
+    assert(typeof options.walletAuth === 'boolean');
+    this.walletAuth = options.walletAuth;
+  }
+
+  if (options.noAuth != null) {
+    assert(typeof options.noAuth === 'boolean');
+    this.noAuth = options.noAuth;
+  }
+
+  if (options.ssl != null) {
+    assert(typeof options.ssl === 'boolean');
+    this.ssl = options.ssl;
+  }
+
+  if (options.host != null) {
+    assert(typeof options.host === 'string');
+    this.host = options.host;
+  }
+
+  if (options.port != null) {
+    assert(typeof options.port === 'number');
+    this.port = options.port;
   }
 
   return this;
